Refactor docs demo architecture: DemoRunner + per-demo files
Introduce shared DemoRunner class and split all demo code into
individual files, eliminating duplication between the homepage
and examples page.

Changes:
- docs/demo-runner.js (new): DemoRunner class handling canvas lifecycle,
  rAF loop, FPS/step stats, 2D/3D rendering, and pointer events.
  Exports addWalls(), spawnRandomShape(), highlightCode(), loadThree().
- docs/demos/*.js (25 new files): each demo in its own module.
  Featured demos (featured:true + featuredOrder 0-8) appear on homepage;
  all 25 appear in the examples grid.
- docs/app.js: 2670 → 395 lines. Imports DemoRunner + demo modules.
  Tab list built dynamically from FEATURED. CodePen/Copy/Benchmark
  unchanged. debugDraw checkbox wired to runner.debugDraw.
- docs/examples.js: 1416 → 203 lines. Each card gets its own DemoRunner.
  Play overlay (▶) replaces auto-start. Per-card stats row (FPS/Step/
  Bodies). "{ } Code" toggle button shows syntax-highlighted code2d.
  IntersectionObserver only pauses already-started demos.
- docs/style.css: play overlay, .card-stats, .card-title-row,
  .code-toggle-btn, .card-code-panel styles added.
- PLAN.md: translated to English.

Duplications eliminated:
  addWalls ×2 → 1, spawnRandomShape ×2 → 1, Three.js setup ×2 → 1,
  pointer event code ×2 → 1, strandbeast demo ×2 → 1.

All 2269 unit tests pass.
